Improvements for caching files in parallel sessions (#5)
- hope this solves the problem for concurrent threads or processes trying to access the cache for the same resource

+++ src/pybiocfilecache/BiocFileCache.py
@@ -2,9 +2,9 @@
 Python Implementation of BiocFileCache
 """
 
-import logging
 import os
 from pathlib import Path
+from time import sleep, time
 from typing import List, Optional, Union
 
 from sqlalchemy import func
@@ -13,12 +13,11 @@ from sqlalchemy.orm import Session
 from .db import create_schema
 from .db.schema import Resource
 from .utils import copy_or_move, create_tmp_dir, generate_id
+from ._exceptions import NoFpathError, RnameExistsError, RpathTimeoutError
 
 __author__ = "jkanche"
 __copyright__ = "jkanche"
 __license__ = "MIT"
-
-_logger = logging.getLogger(__name__)
 
 
 class BiocFileCache:
@@ -30,7 +29,7 @@ class BiocFileCache:
         """Initialize BiocFileCache.
 
         Args:
-            cacheDirOrPath (Union[str, Path], optional): Path to cache
+            cacheDirOrPath (Union[str, Path], optional): Path to cache.
                 directory. Defaults to tmp location, `create_tmp_dir()`.
 
         Raises:
@@ -65,42 +64,56 @@ class BiocFileCache:
         Args:
             rname (str): Name of the resource to add to cache.
             fpath (Union[str, Path]): Location of the resource.
-            rtype (str, optional): One of `local`, `web`, or `relative`.
+            rtype (str, optional): One of `"local"`, `"web"`, or `"relative"`.
                 Defaults to `"local"`.
-            action (str, optional): Either `copy`, `move` or `asis`.
+            action (str, optional): Either `"copy"`, `"move"` or `"asis"`.
                 Defaults to `"copy"`.
             ext (bool, optional): Use filepath extension when storing in cache.
                 Defaults to `False`.
 
         Returns:
-            Resource: Database record of the new resource in cache
-        """
+            Resource: Database record of the new resource in cache.
 
+        Raises:
+            NoFpathError: When the `fpath` does not exist.
+            RnameExistsError: When the `rname` already exists in the cache.
+            sqlalchemy exceptions: When something is up with the cache.
+        """
         if isinstance(fpath, str):
             fpath = Path(fpath)
 
         if not fpath.exists():
-            raise Exception(f"Resource at {fpath} does not exist.")
-
-        existing_cache = self.get(rname)
-        if existing_cache is not None:
-            _logger.info(f"File with {rname} already exists, updating instead")
-            return self.update(rname, fpath, action)
+            raise NoFpathError(f"Resource at {fpath} does not exist.")
 
         rid = generate_id()
-        rpath = f"{self.cache}/{rid}.{fpath.suffix}" if ext else f"{self.cache}/{rid}"
-
-        copy_or_move(str(fpath), rpath, rname, action)
+        rpath = f"{self.cache}/{rid}" + (f".{fpath.suffix}" if ext else "")
 
         # create new record in the database
         res = Resource(
-            **dict(rid=rid, rname=rname, rpath=rpath, rtype=rtype, fpath=str(fpath))
+            **dict(rid=rid, rname=rname, rpath=rpath, rtype=rtype, fpath=str(fpath),)
         )
+
+        # If this was higher up a parallel process could have added the key to
+        # the cache in the meantime as the above takes a bit, so checking here
+        # reduces the odds of this happening
+        # Redirecting to update was removed as it is a scenario better handled
+        # by the caller.
+        if self.get(rname) is not None:
+            raise RnameExistsError("Resource already exists in cache!")
 
         with self.sessionLocal() as session:
             session.add(res)
             session.commit()
             session.refresh(res)
+
+        # In the "move" scenario if we move the file to rpath before rpath is
+        # part of the cache and then when trying to add it to the cache an
+        # exception is raised (such as if it is locked by another process) the
+        # data essentially disappears to rpath with no way of retrieving its
+        # location. Thus we add rpath to the cache first, then move the data
+        # into it so that the data at source does not disappear if accessing
+        # the cache raises an exception.
+        copy_or_move(str(fpath), rpath, rname, action)
 
         return res
 
@@ -108,11 +121,11 @@ class BiocFileCache:
         """Search cache for a resource.
 
         Args:
-            query (str): query to search
+            query (str): query or keywords to search.
             field (str, optional): Field to search. Defaults to "rname".
 
         Returns:
-            List[Resource]: list of matching resources from cache
+            List[Resource]: list of matching resources from cache.
         """
         with self.sessionLocal() as session:
             return (
@@ -129,15 +142,32 @@ class BiocFileCache:
             rname (str): The `rname` of the `Resource` to get.
 
         Returns:
-            res (Resource, optional): The `Resource` for the `rname` if any.
+            (Resource, optional): The `Resource` for the `rname` if any.
         """
-        return session.query(Resource).filter(Resource.rname == rname).first()
+        resource: Optional[Resource] = (
+            session.query(Resource).filter(Resource.rname == rname).first()
+        )
+
+        if resource is not None:
+            # `Resource` may exist but `rpath` could still be being
+            # moved/copied into by `add`, wait until `rpath` exists
+            start = time()
+            timeout = 30
+            while not Path(str(resource.rpath)).exists():
+                if time() - start >= timeout:
+                    raise RpathTimeoutError(
+                        f"For resource: '{rname}' the rpath does not exist "
+                        f"after {timeout} seconds."
+                    )
+                sleep(0.1)
+
+        return resource
 
     def get(self, rname: str) -> Optional[Resource]:
         """get resource by name from cache.
 
         Args:
-            rname (str): Name of the file to search
+            rname (str): Name of the file to search.
 
         Returns:
             Optional[Resource]: matched resource from cache if exists.
@@ -148,14 +178,16 @@ class BiocFileCache:
         """Remove a resource from cache by name.
 
         Args:
-            rname (str): Name of the resource to remove
+            rname (str): Name of the resource to remove.
         """
         with self.sessionLocal() as session:
-            res: Resource = self._get(session, rname)
-            session.delete(res)
-            session.commit()
-            # remove file
-            Path(res.rpath).unlink()
+            res: Optional[Resource] = self._get(session, rname)
+
+            if res is not None:
+                session.delete(res)
+                session.commit()
+                # remove file
+                Path(res.rpath).unlink()
 
     def purge(self):
         """Remove all files from cache."""
@@ -168,12 +200,12 @@ class BiocFileCache:
         """Update a resource in cache.
 
         Args:
-            rname (str): name of the resource in cache
+            rname (str): name of the resource in cache.
             fpath (Union[str, Path]): new resource to replace existing file in cache.
-            action (str, optional): either copy of move. defaults to copy
+            action (str, optional): either copy of move. defaults to copy.
 
         Returns:
-            Resource: Updated resource record in cache
+            Resource: Updated resource record in cache.
         """
 
         if isinstance(fpath, str):
@@ -183,13 +215,18 @@ class BiocFileCache:
             raise Exception(f"File: '{fpath}' does not exist")
 
         with self.sessionLocal() as session:
-            res: Resource = self._get(session, rname)
-            # copy the file to cache
-            copy_or_move(str(fpath), str(res.rpath), rname, action)
-            # TODO: is this needed?
-            res.create_time = res.access_time = res.last_modified_time = func.now()
-            session.merge(res)
-            session.commit()
-            session.refresh(res)
+            res = self._get(session, rname)
 
-        return res
+            if res is not None:
+                # copy the file to cache
+                copy_or_move(str(fpath), str(res.rpath), rname, action)
+                res.access_time = res.last_modified_time = func.now()
+                session.merge(res)
+                session.commit()
+                session.refresh(res)
+            else:
+                # technically an error since update shouldn't be called on
+                # non-existent resources in cache.
+                # but lets just add it to the cache.
+                res = self.add(rname=rname, fpath=fpath, action=action)
+            return res
